wip: MAME 0209 update
This is dependent on an update to MAME which provides a minimal API
for hosting in other front-ends. The headless API includes some
minor modifications to MAME allow single-frame execution of a machine
via a simple Objective-C API. Longer term goals include exposing
more of MAME's configuration to OpenEmu.

Follow development of Headless MAME here:

https://github.com/stuartcarnie/mame/tree/mame0209-headless